refactor: change GameMap creation and move Scala2P to package 'model.utils'
