+++ src/main/scala/model/utils/Scala2P.scala
@@ -1,4 +1,4 @@
-package model.player
+package model.utils
 
 /** Object that provides utility methods to set theories and resolve Prolog goals in Scala. Adapted from
   * https://github.com/unibo-pps/pps-lab12/blob/master/src/main/scala/it/unibo/u12lab/code/Scala2P.scala
